factor grpcMock out of doctest-settings
