+++ .markdown-doctest-setup.js
@@ -3,49 +3,26 @@
 const assert = require('assert');
 
 const rchain = require('.');
+const { grpcMock } = require('./test/testRNode');
 const { RNode, RholangCrypto, REV, Ed25519keyPair, Hex } = rchain;
 
 
-// mock enough of grpc
-// ISSUE: refactor overlap with test/testRNode.js
-const CasperMessage = require('./protobuf/CasperMessage');
-const { Either } = CasperMessage;
-const { BlockInfoWithoutTuplespace } = CasperMessage.coop.rchain.casper.protocol;
+function config(env, grpcAccess) {
+  const grpc = env.NODE_ENV === 'production' ? grpcAccess() : grpcMock();
+  console.log('doctest env:', env.NODE_ENV);
 
-function DeployService(_hostPort, _chan) {
-  return Object.freeze({
-    doDeploy(_dd, _auto = false) { return 'Success!'; },
-    showBlocks(_depth) {
-      const block4 = { value: BlockInfoWithoutTuplespace.encode({ blockNumber: 4 }).finish() };
-
-      return Object.freeze({
-        on(name, handler) {
-          if (name === 'data') {
-            handler({ success: { response: block4 } });
-          } else if (name === 'end') {
-            handler();
-          }
-        }
-      });
+  return {
+    babel: false,
+    require: {
+      grpc: grpc,  // ISSUE: live access at test time?
     },
-  });
+    globals: {
+      assert, Buffer,
+      grpc,
+      RholangCrypto, RNode, REV, Hex, Ed25519keyPair,
+    },
+  };
 }
 
-const casper = { DeployService };
-const proto = { coop: { rchain: { casper: { protocol: casper } } } };
-const grpc = {
-  loadPackageDefinition(_d) { return proto; },
-  credentials: { createInsecure() { } },
-};
-
-module.exports = {
-  babel: false,
-  require: {
-    grpc: grpc,  // ISSUE: live access at test time?
-  },
-  globals: {
-    assert, Buffer,
-    grpc,
-    RholangCrypto, RNode, REV, Hex, Ed25519keyPair,
-  },
-};
+/* global process */
+module.exports = config(process.env, () => require('grpc'));
